+++ example/lib/use_cases/installed_apps_page.dart
@@ -57,6 +57,7 @@ class _InstalledAppsPageState extends State<InstalledAppsPage> {
     _InstalledItem('Maps', 'Moovit', Moovit(), 'assets/moovit.png'),
     _InstalledItem('Maps', 'Neshan', Neshan(), 'assets/neshan.png'),
     _InstalledItem('Maps', 'Yandex Maps', YandexMaps(), 'assets/yandex_maps.png'),
+    _InstalledItem('Maps', 'Yandex Navigator', YandexNavigator(), 'assets/yandex_navigator.png'),
     _InstalledItem('Stores', 'Google Play Store', PlayStore(), 'assets/google_play.png'),
     _InstalledItem('Stores', 'iOS App Store', IOSAppStore(), 'assets/apple.png'),
     _InstalledItem('Stores', 'Mac App Store', MacAppStore(), 'assets/apple.png'),