+++ example/test/use_cases_test.dart
@@ -22,7 +22,7 @@ void main() {
       await tester.pumpWidget(const MaterialApp(home: HomePage()));
 
       expect(find.text('View all 9'), findsOneWidget);
-      expect(find.text('View all 22'), findsOneWidget);
+      expect(find.text('View all 23'), findsOneWidget);
       expect(find.text('View all 7'), findsOneWidget);
 
       const titles = [
@@ -44,7 +44,7 @@ void main() {
 
       await tester.tap(find.byKey(const ValueKey('filter-apps')));
       await tester.pump();
-      expect(find.byType(AppTile), findsNWidgets(22));
+      expect(find.byType(AppTile), findsNWidgets(23));
 
       await tester.tap(find.byKey(const ValueKey('filter-stores')));
       await tester.pump();
@@ -207,10 +207,11 @@ void main() {
       await tester.pumpAndSettle();
 
       final actions = deeplinkX.mapDirectionsActions.single;
-      expect(actions, hasLength(10));
+      expect(actions, hasLength(11));
       expect(actions.first, isA<GoogleMapsDirectionsWithCoordsAction>());
       expect(actions[6], isA<SygicDirectionsWithCoordsAction>());
-      expect(actions.last, isA<YandexMapsDirectionsWithCoordsAction>());
+      expect(actions[9], isA<YandexMapsDirectionsWithCoordsAction>());
+      expect(actions.last, isA<YandexNavigatorDirectionsWithCoordsAction>());
     });
 
     testWidgets('rating selector uses Android listing alternatives', (final tester) async {